requirements.txt
tensorflow>=2.12
numpy
pillow
streamlit